refactor(table-advance): remove mock data to appropriate file
